Make clippy -D warnings clean
The CI clippy gate reported 21 warnings. Fix the genuine ones and allow the
C-mirroring idioms:

  * unnecessary_cast (15, e.g. `u8 as u8`, `i32 as i32`) — removed via
    `cargo clippy --fix` across 8 files. These were no-op casts the transpile
    emitted; dropping them is a strict improvement.
  * explicit_counter_loop (4) and manual_checked_ops (2) — allowed in
    Cargo.toml with rationale. These ports mirror tmux's C control flow 1:1
    (explicit index counters, `if d != 0 { n / d }` guards); clippy's rewrites
    would diverge from the C reference we validate against. Revisited per-module
    as code is brought to idiomatic Rust.

`cargo clippy --all-targets --locked -- -D warnings` now passes; the CI clippy
job goes green (it stays advisory per the sibling pattern). Builds and runs.

+++ src/cmd_parse.rs
@@ -912,7 +912,7 @@ unsafe fn yylex_(ps: &mut cmd_parse_state) -> Option<Tok> {
                 let yylval_token = yylex_get_word(ps, '%' as i32);
                 let mut cp = yylval_token;
                 while *cp != b'\0' {
-                    if *cp != b'%' && !(*cp as u8).is_ascii_digit() {
+                    if *cp != b'%' && !(*cp).is_ascii_digit() {
                         break;
                     }
                     cp = cp.add(1);
